描画速度の高速化 画面の範囲外まで描画されていたのを修正 ホールド時のスクロールのスピードを1f分高速に。

+++ src/player.cpp
@@ -97,7 +97,7 @@ double Player::GetAudioPosition(void)
 	{
 		return 0;
 	}
-	return (Cursor / Length);
+	return ((double)Cursor / Length);
 }
